# perf(model): add caching for data keys